feat: add missing eventing roles, clusterroles, and bindings
This currently does not work.  I think the aggregation roles (such as the ClusterRole `addressable-resolver`) fails when doing an `apply`, possibly because it tries to overwrite the `rules` attribute (which will be dynamically set by Kubernetes and should not be set from yaml).  This can be seen in the juju debug logs, where we see an error with message like:

unit-knative-eventing-controller-0: 17:18:00 ERROR unit.knative-eventing-controller/0.juju-log Error installing charm: Apply failed with 1 conflict: conflict with "clusterrole-aggregation-controller": .rules