fix: comment out aggregate Role rules
using lightkube.apply on an aggregated Role or ClusterRole leads to a conflict error on the .rules as described [here](https://github.com/canonical/knative-operators/issues/6).  This fix manually removes the `.rules` to avoid this conflict.

This change solves the above problem with no negative side effect with respect to function (there are no cases where it breaks functionality), but it does require maintaining a difference from the upstream manifest.