fix: rename eventing configmap to config-features
Previously, Eventing's `config-features` configmap was renamed from `config-features` to `eventing-config-features` to avoid a naming collision with Serving's `config-features` file.  This commit reverts that change, resulting in a break where eventing-controller and serving-controller charms cannot be deployed at the same time (because they both try to control the same configmap).

Both eventing and serving rely on a configmap named `config-features` to exist in their namespace.  These configmaps define the feature flags for the respective apps, and each application has different feature flags (eg: each expects different values in `config-features`).  It appears upstream expects eventing and serving to be deployed in separate namespaces in order to avoid collision of these configmaps (and likely other objects).  But knative does not expose an input argument to define the name of the configmap they're looking for, so we cannot change the configmap to `eventing-config-features`