feat: Use experimental apply_many function to apply yaml
Some objects in the yaml manifests for eventing are ordered in a way that can not be sequentially deployed.  For example, we might have:

```
kind: ClusterRoleBinding
roleRef:
  kind: ClusterRole
  name: example-cluster-role-binding
subjects:
- kind: ServiceAccount
  name: example-service-account
...
---
kind: ClusterRole
metadata:
  name: example-cluster-role
...
---
kind: ServiceAccount
metadata:
  name: example-service-account
```

Which defines a ClusterRoleBinding before the ClusterRole and ServiceAccount it acts upon.  This leads to errors when applying the yaml.

apply_many handles this by deploying batches of resources in order by kind, putting things like ClusterRoleBindings after the kinds of ClusterRole or ServiceAccount.