knative-operator/src/charm.py: small refactor
We now have a `_main` method for handling the main changes
in the charm: applying k8s resources and configuring pebble layers.
This commit also adds exception handling and logging for when pebble
throws an APIError.
Finally, restarting the underlying service should be done only if
changes were made, thus moving the restart call inside the statement
that checks for changes.