knative-operator/src/charm.py: refactors update_layer and main
update_layer now uses replan() instead of restart()
for correctly reconciling the workload desired state,
we are also adding error handling for ensuring the correct
status is set if the replan fails.
main now makes calls to event handlers and lets them set the
status and do the error handling.