fix: change how we check if knative service is available
Previously, we listedon KNative Services to confirm that they were available for use.  This test might have been failing because we had not specified the namespace to list under, leading to us using a default namespace that may not exist.  This updated version refactors the check, instead looking directly for the expected CRD rather than just listing those CRDs.  This makes the check more directly test the thing we care about.