feat: add library for sharing the localGateway data
knative-serving is responsible for configuring and creating its localGateway,
an internal ingress Gateway that knative uses for handling requests.
This localGateway is often used by other components that integrate with Knative
for serverless or event driven operations, for example Kserve in Serverless mode.
Since the localGateway is configured by the knative-serving charm using configuration
values that are unknown to other charms, a library and a relation are needed for exposing them.