feat: enable support for Gateway APIs in Knative Serving
This commit adds basic support for Gateway APIs in Knative Serving.  Knative provides a net-gateway-api controller to support the Gateway APIs, but it is not fully supported by Knative's Operator.  This commit thus:

* adds charm config for the `ingress-class`, which can optionally be set to `gateway-api.ingress.networking.knative.dev` to enable the Gateway API mode
* if `ingress-class=gateway-api...`, this commit deploys the net-gateway-api controller and configures it based on charm config

While we deploy and configure net-gateway-api when ingress-class is set to gateway-api, we do not actually remove the net-istio controller from the cluster.  Instead, net-istio is still deployed (but sits idle).  Not sure how to avoid installing it.